requirements.txt
ultralytics>=8.2.0
opencv-contrib-python>=4.9.0.80
numpy>=1.26.0
Flask>=3.0.0
timm>=0.9.12
openai>=1.13.0
requests>=2.31.0
python-dotenv>=1.0.1
huggingface_hub>=0.23.0
